fix frontend merge issues
