+++ SourceCode_and_Documentation/backend/README.md
@@ -39,4 +39,5 @@ uvicorn main:app --reload
 
 Note, the reload flag makes it development only, might add in a deployment script at some point
 
-## Backend Directory Structure (to be added)
+## Backend Directory Structure
+For a detailed repository structure check out [RepositoryStructure.md](RepositoryStructure.md)
